feat(playground): simplify embed logic
