+++ packages/playground/src/contexts/index.ts
@@ -2,5 +2,4 @@ import { ChatContext } from "./chat-context";
 import { RootContext } from "./root-context";
 
 export { ChatContext, RootContext };
-export { EmbedPreloadContext, useEmbedPreload } from "./embed-preload-context";
 export { NavbarContext, NavbarProvider } from "./navbar-context";
