+++ packages/playground/src/pages/main-layout.tsx
@@ -1,12 +1,6 @@
 import { observer } from "mobx-react-lite";
 import React, { type ReactNode, useEffect, useMemo, useState } from "react";
-import {
-	Link,
-	matchPath,
-	Outlet,
-	useLocation,
-	useNavigate,
-} from "react-router-dom";
+import { Link, matchPath, Outlet, useLocation } from "react-router-dom";
 import { useInsight } from "@semoss/sdk/react";
 import {
 	Breadcrumb,
@@ -22,9 +16,8 @@ import {
 } from "@semoss/ui/next";
 import { GlobalFooter, GlobalNav } from "@/components";
 import { GlobalDialog } from "@/components/common/global-dialog";
-import { ChatContext, EmbedPreloadContext, NavbarContext } from "@/contexts";
+import { ChatContext, NavbarContext } from "@/contexts";
 import { useRoot } from "@/hooks";
-import { useNavbar } from "@/hooks/use-navbar";
 import { useThemeTitle } from "@/hooks/use-theme-title";
 import { ChatStore } from "@/stores";
 import { setFavicon } from "@/utility/utils";
@@ -34,13 +27,14 @@ export const MainLayout = observer(() => {
 	const { root } = useRoot();
 	const theme = root.theme;
 	const [navbarActions, setNavbarActions] = useState<ReactNode | null>(null);
+	const { pathname } = useLocation();
 
 	const [isSidebarOpen, setIsSidebarOpen] = useCacheState(
 		false,
 		`sidebar--isOpen`,
 	);
 
-	// set up the store
+	// set up the chat store
 	const chatStore = useMemo(() => {
 		const store = new ChatStore(
 			root.theme,
@@ -70,86 +64,6 @@ export const MainLayout = observer(() => {
 			<NavbarContext.Provider
 				value={{ actions: navbarActions, setActions: setNavbarActions }}
 			>
-				<MainLayoutContent
-					isSidebarOpen={isSidebarOpen}
-					setIsSidebarOpen={setIsSidebarOpen}
-				/>
-			</NavbarContext.Provider>
-		</ChatContext.Provider>
-	);
-});
-
-const MainLayoutContent = observer(
-	({
-		isSidebarOpen,
-		setIsSidebarOpen,
-	}: {
-		isSidebarOpen: boolean;
-		setIsSidebarOpen: (open: boolean) => void;
-	}) => {
-		const { root } = useRoot();
-		const { actions } = useNavbar();
-		const navigate = useNavigate();
-		const { pathname } = useLocation();
-
-		// Collect sidebar items that should be embedded as iframes
-		const embedItems = useMemo(
-			() =>
-				[
-					...(root.theme.sidebar.headerItems ?? []),
-					...(root.theme.sidebar.footerItems ?? []),
-				].filter((item) => item.embed && item.url),
-			// eslint-disable-next-line react-hooks/exhaustive-deps
-			[root.theme.sidebar.headerItems, root.theme.sidebar.footerItems],
-		);
-
-		// Map of path → src; derived immediately so iframes start loading right away.
-		const preloadedSrcs = useMemo(
-			() =>
-				Object.fromEntries(
-					embedItems.map((item) => [item.path, item.url]),
-				),
-			[embedItems],
-		);
-
-		useEffect(() => {
-			const handleMessage = (event: MessageEvent) => {
-				if (event.data?.type === "SMSS_NEW_CHAT") {
-					const { workspaceId, knowledgeId } =
-						event.data.payload ?? {};
-					if (workspaceId) {
-						navigate(`/new?workspaceId=${workspaceId}`);
-					} else if (knowledgeId) {
-						navigate(`/new?knowledgeId=${knowledgeId}`);
-					} else {
-						navigate(`/new`);
-					}
-				} else if (event.data?.type === "SMSS_OPEN_ROOM") {
-					const { roomId, jobId, prompt } = event.data.payload ?? {};
-					if (roomId) {
-						const params = new URLSearchParams();
-						if (jobId) params.set("jobId", jobId);
-						if (prompt) params.set("prompt", prompt);
-						const qs = params.toString();
-						navigate(
-							qs ? `/room/${roomId}?${qs}` : `/room/${roomId}`,
-						);
-					}
-				}
-			};
-			window.addEventListener("message", handleMessage);
-			return () => window.removeEventListener("message", handleMessage);
-		}, [navigate]);
-
-		// The set of paths that have been pre-loaded; shared with EmbedPage so
-		// it can skip rendering its own duplicate iframe.
-		const preloadedPaths = useMemo(
-			() => new Set(Object.keys(preloadedSrcs)),
-			[preloadedSrcs],
-		);
-
-		return (
-			<EmbedPreloadContext.Provider value={preloadedPaths}>
 				<SidebarProvider
 					open={isSidebarOpen}
 					onOpenChange={setIsSidebarOpen}
@@ -223,25 +137,24 @@ const MainLayoutContent = observer(
 								</div>
 								<div className="flex-1" />
 								<div className="flex items-center gap-2">
-									{actions ?? null}
+									{navbarActions ?? null}
 								</div>
 							</div>
 							<Separator />
 							<div className="relative w-full flex-1 overflow-hidden">
 								<Outlet />
-								{embedItems.map((item) => {
-									const isActive = !!matchPath(
-										`/embed/${item.path}`,
+								{Object.entries(
+									chatStore.preloadedEmbedPathMap,
+								).map(([path, url]) => {
+									const isActive = matchPath(
+										`/embed/${path}`,
 										pathname,
 									);
 									return (
 										<iframe
-											key={item.path}
-											src={
-												preloadedSrcs[item.path] ??
-												undefined
-											}
-											title={item.name}
+											key={path}
+											src={url}
+											title={path}
 											className="absolute inset-0 h-full w-full border-none"
 											// @ts-expect-error fetchpriority is not yet in React's typings
 											fetchpriority="high"
@@ -259,7 +172,7 @@ const MainLayoutContent = observer(
 						</div>
 					</SidebarInset>
 				</SidebarProvider>
-			</EmbedPreloadContext.Provider>
-		);
-	},
-);
+			</NavbarContext.Provider>
+		</ChatContext.Provider>
+	);
+});
